Fixed images with no links bug
d) The search works correctly when a book does not have a thumbnail or an author. (To test this, try searching for "poetry" and "biography"). Some books do not have an imageLinks property, this is the case when a user enter biography as the term, try checking first if the book have this property before you use imageLinks.smallThumbnails